transaction-parsing-service/requirements.txt
pandas
psycopg2-binary
structlog
python-dotenv
base58